requirements.txt
ccxt==4.2.85
pandas==2.1.4
plotly==5.18.0
pyyaml==6.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
numpy==1.26.4